# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for go-api.",
        "title": "go-api API",
        "contact": {
            "name": "Lewis",
            "email": "lxx0103@yahoo.com"
        },
        "version": "1.0"
    },
    "host": "0.0.0.0:8080",
    "basePath": "/",
    "paths": {
...
        "/purchaseorders": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "采购单管理"
                ],
                "summary": "采购单列表",
                "operationId": "402",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "每页行数（5/10/15/20）",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "采购单号码",
                        "name": "purchaseorder_number",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ListRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/purchaseorder.PurchaseorderResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "采购单管理"
                ],
                "summary": "新建采购单",
                "operationId": "401",
                "parameters": [
                    {
                        "description": "采购单信息",
                        "name": "purchaseorder_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/purchaseorder.PurchaseorderNew"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            }
        },
        "/purchaseorders/:id": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "采购单管理"
                ],
                "summary": "根据ID获取采购单",
                "operationId": "404",
                "parameters": [
                    {
                        "type": "string",
                        "description": "采购单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/purchaseorder.PurchaseorderResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "采购单管理"
                ],
                "summary": "根据ID更新采购单",
                "operationId": "403",
                "parameters": [
                    {
                        "type": "string",
                        "description": "采购单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "采购单信息",
                        "name": "purchaseorder_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/purchaseorder.PurchaseorderNew"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/purchaseorder.Purchaseorder"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "采购单管理"
                ],
                "summary": "根据ID删除采购单",
                "operationId": "405",
                "parameters": [
                    {
                        "type": "string",
                        "description": "采购单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            }
        },
        "/purchaseorders/:id/issued": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "采购单管理"
                ],
                "summary": "更新采购单为ISSUED",
                "operationId": "407",
                "parameters": [
                    {
                        "type": "string",
                        "description": "采购单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ListRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            }
        },
        "/purchaseorders/:id/items": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "采购单管理"
                ],
                "summary": "采购单产品列表",
                "operationId": "406",
                "parameters": [
                    {
                        "type": "string",
                        "description": "采购单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ListRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/purchaseorder.PurchaseorderItemResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            }
        },
        "/purchaseorders/:id/receives": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "采购单管理"
                ],
                "summary": "新建收货单",
                "operationId": "408",
                "parameters": [
                    {
                        "description": "采购单信息",
                        "name": "purchasereceive_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/purchaseorder.PurchasereceiveNew"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            }
        },
        "/purchasereceives": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "采购单管理"
                ],
                "summary": "收货单列表",
                "operationId": "409",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "每页行数（5/10/15/20）",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "采购单号码",
                        "name": "purchasereceive_number",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ListRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/purchaseorder.PurchasereceiveResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            }
        },
        "/purchasereceives/:id/details": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "采购单管理"
                ],
                "summary": "收货单详情列表",
                "operationId": "411",
                "parameters": [
                    {
                        "type": "string",
                        "description": "收货单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ListRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/purchaseorder.PurchasereceiveDetailResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            }
        },
        "/purchasereceives/:id/items": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "采购单管理"
                ],
                "summary": "收货单商品列表",
                "operationId": "410",
                "parameters": [
                    {
                        "type": "string",
                        "description": "收货单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ListRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/purchaseorder.PurchasereceiveItemResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "purchaseorder.PurchaseorderResponse": {
            "type": "object",
            "properties": {
                "billing_status": {
                    "type": "integer"
                },
                "discount_type": {
                    "type": "integer"
                },
                "discount_value": {
                    "type": "number"
                },
                "expected_delivery_date": {
                    "type": "string"
                },
                "item_count": {
                    "type": "number"
                },
                "notes": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "string"
                },
                "purchaseorder_date": {
                    "type": "string"
                },
                "purchaseorder_id": {
                    "type": "string"
                },
                "purchaseorder_number": {
                    "type": "string"
                },
                "receive_status": {
                    "type": "integer"
                },
                "shipping_fee": {
                    "type": "number"
                },
                "status": {
                    "type": "integer"
                },
                "sub_total": {
                    "type": "number"
                },
                "tax_total": {
                    "type": "number"
                },
                "total": {
                    "type": "number"
                },
                "vendor_id": {
                    "type": "string"
                },
                "vendor_name": {
                    "type": "string"
                }
            }
        },
...
        "purchaseorder.PurchasereceiveItemNew": {
            "type": "object",
            "required": [
                "quantity"
            ],
            "properties": {
                "item_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "purchaseorder.PurchasereceiveNew": {
            "type": "object",
            "required": [
                "items",
                "purchasereceive_date",
                "purchasereceive_number"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/purchaseorder.PurchasereceiveItemNew"
                    }
                },
                "notes": {
                    "type": "string"
                },
                "purchasereceive_date": {
                    "type": "string"
                },
                "purchasereceive_number": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 6
                }
            }
        },
...
        "response.ErrorRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ListRes": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "data": {},
                "page_id": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
        "response.SuccessRes": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
...
    }
}